+++ frontend/src/pages/InterventionForm/InterventionFormPage.js
@@ -5,7 +5,11 @@ import styles from "./InterventionFormPage.module.css";
 import stylesMobile from "./InterventionFormPageMobile.module.css";
 import TemplateGlobal from "../Template/TemplateGlobal";
 import TemplateGlobalMobile from "../Template/TemplateGlobalMobile";
-import { createReport, createEvent } from "../../services/api";
+import {
+    createReport,
+    createEvent,
+    createNotification,
+} from "../../services/api";
 import Canvas from "../../components/Canvas/Canvas";
 
 // Composant wrapper pour utiliser les hooks
@@ -163,6 +167,27 @@ class InterventionFormPage extends Component {
                 }
             });
         }
+
+        const notificationData = {
+            idEmployee: eventDetails.employee.idEmployee,
+            action: "Validation",
+            type: eventDetails.type,
+            title: eventDetails.title,
+            creationDate: new Date().toISOString().split("T")[0],
+            creationHour: new Date().toLocaleTimeString("fr-FR", {
+                hour: "2-digit",
+                minute: "2-digit",
+                second: "2-digit",
+            }),
+        };
+
+        createNotification(notificationData, (error, result) => {
+            if (error) {
+                console.error("Error creating notification:", error);
+            } else {
+                console.log("Notification created successfully:", result);
+            }
+        });
 
         const reportData = {
             breakdown,