+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Changes
 
 -   Add option to create an unscheduled intervention when the technician validates the intervention form
+-   Add notifications for creating, updating, deleting, and validating events with employee and date details
 
 ## Fixes